Reorganise pixel access functions into separate modules Add docstrings for pixel_buffer functions Fix non-contiguous pixel buffers used with set_pixels_matrix_hack Broadcast color squares to the desired shape instead of creating full width*height arrays of the same color Explicitly use fallback get_pixels_no_gl when gl_load fails Fix pixel_buffer paste fitting oversize images 1 pixel out of bounds
